Osa4 4.23 final. Testien korjaus ja lisäys autentikoinnin osalta

+++ osa4/blogilista/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -25,12 +23,7 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const blog = await new Blog(request.body)
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   if (!blog){
     response.status(400).json({ error: 'userId missing or not valid' })
@@ -45,27 +38,47 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!blog) {
+    response.status(404).json({ error: 'blog not found' })
+  } else if (!blog.user){
+    response.status(403).json({ error: 'there\'s no key `user` in this blog' })
+  }
+
+  if ( blog.user.toString() === user._id.toString() ){
+    await Blog.findByIdAndDelete(request.params.id)
+  } else {
+    return response.status(401).json({ error: 'unauthorized' })
+  }
+
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
   if (Object.keys(request.body).includes('_id')){
     response
       .status(400)
       .send('DB _id is not allowed to change')
       .end()
+  } else if (!blog.user){
+    response.status(403).json({ error: 'there\'s no key `user` in this blog' })
+  } else if ( blog.user.toString() === user._id.toString() ){
+    await Blog
+      .findByIdAndUpdate(request.params.id, request.body, {
+        new: true,
+        runValidators: true
+      })
+    response
+      .status(200)
+      .end()
+  } else {
+    response.status(401).json({ error: 'unauthorized' })
   }
-
-  await Blog
-    .findByIdAndUpdate(request.params.id, request.body, {
-      new: true,
-      runValidators: true
-    })
-
-  response
-    .status(200)
-    .end()
 })
 
 module.exports = blogsRouter
